Handle possible errors during build

+++ build.ps1
@@ -1,7 +1,8 @@
 param(
     [switch] $CleanDist = $true,
     [switch] $RebuildShell = $false,
-    [switch] $RebuildBuildImages = $false
+    [switch] $RebuildBuildImages = $false,
+    [switch] $RebuildFailed = $false
 )
 # This variable for using different .dockerignore files for different Dockerfiles
 $env:DOCKER_BUILDKIT=1
@@ -10,6 +11,15 @@ $env:DOCKER_BUILDKIT=1
 $NgShell = 'iqs-libs-clientshell-angular'
 $NgImageExists = docker images ${NgImage} -q
 $NgxImageExists = docker images ${NgxImage} -q
+$global:iqsBuildErrors = @()
+$Sources = $Dists
+if (Test-Path -Path ".\.iqsBuildErrors.json") {
+    if ($RebuildFailed) {
+        $CleanDist = $false
+        $Sources = Get-Content ./.iqsBuildErrors.json | ConvertFrom-Json
+    }
+    Remove-Item ./.iqsBuildErrors.json -Force
+}
 
 if (($NgImageExists -and $RebuildBuildImages) -or (-not $NgImageExists)) {
     if ($NgImageExists) {
@@ -43,14 +53,21 @@ function Ng-Build {
         docker cp ../${NgShell}/dist ${NgContainer}:/usr/src/app/${NgShell}/dist
         docker cp ../iqs-libs-unsupported-angular ${NgContainer}:/usr/src/app
     }
-    docker exec ${NgContainer} /bin/sh -c "cd ${Dir} && gulp build"
-    if (-Not $IsShell) {
-        docker cp ${NgContainer}:/usr/src/app/${Dir}/dist ../dist/${Dist}
+    docker exec ${NgContainer} /bin/sh -c "cd ${Dir} && gulp build 2>&1 | tee /var/log/gulp.log"
+    if (docker exec ${NgContainer} /bin/sh -c "grep -r 'error' /var/log/gulp.log") {
+        Write-Host "Error during build: ${Dir}" -ForegroundColor White -BackgroundColor Red
+        New-Item -itemtype directory "../.logs/$Dist" -Force
+        docker cp ${NgContainer}:"/var/log/gulp.log" "../.logs/$Dist/"
+        $global:iqsBuildErrors += @{ dir = $Dir; type = 'ng'; dist = $Dist }
     } else {
-        if (Test-Path -Path ".\dist") {
-            Remove-Item -LiteralPath "dist" -Recurse -Force
+        if (-Not $IsShell) {
+            docker cp ${NgContainer}:/usr/src/app/${Dir}/dist ../dist/${Dist}
+        } else {
+            if (Test-Path -Path ".\dist") {
+                Remove-Item -LiteralPath "dist" -Recurse -Force
+            }
+            docker cp ${NgContainer}:/usr/src/app/${Dir}/dist .
         }
-        docker cp ${NgContainer}:/usr/src/app/${Dir}/dist .
     }
     docker stop $NgContainer
     docker rm $NgContainer
@@ -74,7 +91,14 @@ function Ngx-Build {
     Remove-Item -LiteralPath "${Dir}.tar"
     docker exec ${NgxContainer} /bin/sh -c "tar -xf ${Dir}.tar"
     docker exec ${NgxContainer} /bin/sh -c "npm i && npm run build:prod"
-    docker cp ${NgxContainer}:/usr/src/app/dist/${Dir} ../dist/${Dist}
+    if (docker exec ${NgxContainer} /bin/sh -c "cd /root/.npm/_logs && ls") {
+        Write-Host "Error during build: ${Dir}" -ForegroundColor White -BackgroundColor Red
+        New-Item -itemtype directory "../.logs/$Dist" -Force
+        docker cp ${NgxContainer}:"/root/.npm/_logs/" "../.logs/$Dist/"
+        $global:iqsBuildErrors += @{ dir = $Dir; type = 'ngx'; dist = $Dist }
+    } else {
+        docker cp ${NgxContainer}:/usr/src/app/dist/${Dir} ../dist/${Dist}
+    }
     docker stop $NgxContainer
     docker rm $NgxContainer
     cd ..
@@ -95,7 +119,7 @@ if (!(Test-Path -Path ".\dist\index.html" -PathType Leaf)) {
     Set-Content .\dist\index.html '<head><meta http-equiv="refresh" content="1;URL=/home/index.html" /></head>'
 }
 
-$Dists | ForEach-Object {
+$Sources | ForEach-Object {
     if ($_.type -eq 'ng') {
         Ng-Build -Dir:$_.dir -Dist:$_.dist
     } else {
@@ -103,7 +127,12 @@ $Dists | ForEach-Object {
     }
 }
 
-if (docker images ${ClientsImage} -q) {
-    docker rmi ${ClientsImage} --force
+if ($global:iqsBuildErrors.Count) {
+    Write-Host "Some errors happened during build process. Run './build.ps1 -RebuildFailed' to rebuild only failed clients"
+    ConvertTo-Json $global:iqsBuildErrors | Out-File -FilePath ./.iqsBuildErrors.json -Force
+} else {
+    if (docker images ${ClientsImage} -q) {
+        docker rmi ${ClientsImage} --force
+    }
+    docker build --no-cache -f ./docker/Nginx.Dockerfile -t ${ClientsImage} .
 }
-docker build --no-cache -f ./docker/Nginx.Dockerfile -t ${ClientsImage} .
